// honey-paradise-client/src/shared/types/server/getMeResponseGender.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Honey Paradise
 * Honey Paradise API
 * OpenAPI spec version: 0.0.2
 */

export type GetMeResponseGender = typeof GetMeResponseGender[keyof typeof GetMeResponseGender];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetMeResponseGender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  OTHER: 'OTHER',
} as const;