// honey-paradise-client/src/shared/types/server/updateUserDto.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Honey Paradise
 * Honey Paradise API
 * OpenAPI spec version: 0.0.2
 */
import type { UpdateUserDtoGender } from './updateUserDtoGender';

export interface UpdateUserDto {
  phoneNumber?: string;
  username?: string;
  gender?: UpdateUserDtoGender;
  birthdate?: string;
}